Added collider objects.
Added line, aabb, and circle classes to use for collisions later.

+++ math_2019/base_conversion.cpp
@@ -5,7 +5,7 @@
 int base_conversion::binary_string_to_integer(const char * a_binary_string)
 {
 	int bit_field = 0;
-	int string_length = strlen(a_binary_string);
+	size_t string_length = strlen(a_binary_string);
 	
 	for (size_t i = 0; i < string_length; i++)
 	{
